requirements.txt
rpaframework==22.1.1
rpaframework-pdf==7.1.3
